// api/src/database/database.types.ts
export type Decimal = ColumnType<string, number | string>;

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export interface Tbclases {
  activa: number;
  clase: number;
  cntlimite: number;
  cobinsc: number;
  controlhr: number;
  descanual: Decimal;
  descmes: Decimal;
  descqna: Decimal;
  descsem: Decimal;
  descstre: Decimal;
  desctrim: Decimal;
  envia: number;
  fecmod: Date;
  fecnvo: Date;
  id: Generated<number>;
  impticketasist: number;
  limitectes: number;
  nomclase: string;
  pranual: Decimal;
  prinsc: Decimal;
  prmes: Decimal;
  prqna: Decimal;
  prsem: Decimal;
  prstre: Decimal;
  prtrim: Decimal;
  usumod: number;
  usunvo: number;
}